+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 ignore = E203, E266, E501, W503, F403, F401
-max-line-length = 89
+max-line-length = 121
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
 exclude=
