+++ xca/examples/arxiv200800283/simple_example.py
@@ -1,11 +1,7 @@
 from xca.examples.arxiv200800283.example_synthesis import pattern_simulation
-from xca.ml.tf_data_proc import xr_dir_to_record, _int64_feature
-from xca.ml.tf_models import (
-    model_training,
-    build_fusion_ensemble_model,
-    build_CNN_model,
-)
-from xca.ml.tf_parameters import load_hyperparameters
+from xca.ml.tf.data_proc import xr_dir_to_record, _int64_feature
+from xca.ml.tf.cnn import build_CNN_model, build_fusion_ensemble_model, training
+from xca.ml.tf.utils import load_hyperparameters
 from pathlib import Path
 
 
@@ -22,7 +18,7 @@ def main():
         model = build_fusion_ensemble_model(
             params.pop("ensemble_size", 1), build_CNN_model, **params
         )
-        res = model_training(model, **params)
+        res = training(model, **params)
         print(f"Results for {system}")
         print(res)
         model.save(str(Path("tmp") / f"{system}"))
